refactor: change namespace of Log
