+++ src/CuiLib/Logging/Logger.cs
@@ -6,7 +6,7 @@ using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace CuiLib.Log
+namespace CuiLib.Logging
 {
     /// <summary>
     /// ロガーのクラスです。
